% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WSD_Function.R
\name{WSD}
\alias{WSD}
\title{Weighted Spectral Difference}
\usage{
WSD(
  data,
  control_temp,
  control_pH,
  units = NULL,
  min_wavelength,
  max_wavelength
)
}
\arguments{
\item{data}{A tidy 5 column dataframe containing wavelength, temperature, Molar Elipticity or Radians Measures from Circular Dichroism Experiment, pH, and protein ID columns}

\item{control_temp}{Temperature of control condition from which to calculate weighted spectral difference}

\item{control_pH}{pH of control condition from which to calculate weighted spectral difference}

\item{units}{(unsupported) The units of input data. Defaults to Molar Elipticity. Options: "MolElip", "Radians", etc}

\item{min_wavelength}{lowest wavelength (from first column of dataframe) to include in weighted spectral difference calculation}

\item{max_wavelength}{highest wavelength (from first column of dataframe) to include in weighted spectral difference calculation}
}
\value{
A tidy data.frame that contains WSD estimates for each condition tested
}
\description{
Input data must be in tidy format.
This function calculates Weighted Spectral Difference for preprocessed data
that has been made with WSD_dataprep or otherwise tidy preprocessed data.
Output is a tidy data.frame with all WSD (and SD) values for input raw data.
}
